ajustes ao updateSaque em serviceUser(BigDecimal -> float) +
remoção de if desnecessario e correção ao metodo de saque.

+++ src/main/java/br/com/lrz/betRaLa/models/User.java
@@ -37,7 +37,7 @@ public class User {
     @Column(nullable = false)
     private String nome; 
     
-    private BigDecimal saldo;
+    private Float saldo;
     
     @Column(nullable = false, unique = true)
     private String email;
